Corrected directory structure, fixed minor details
