+++ include/magma/Buffer.hpp
@@ -18,7 +18,7 @@ namespace magma
   template<class Deleter = BufferDeleter>
   using Buffer = claws::handle<vk::Buffer, Deleter>;
 
-  inline auto DeviceImpl::createBuffer(vk::BufferCreateFlags flags,
+  inline auto impl::Device::createBuffer(vk::BufferCreateFlags flags,
                                        vk::DeviceSize size,
                                        vk::BufferUsageFlags usage,
                                        std::vector<uint32_t> const &queueFamilies) const
@@ -28,7 +28,7 @@ namespace magma
                       {flags, size, usage, vk::SharingMode::eExclusive, static_cast<uint32_t>(queueFamilies.size()), queueFamilies.data()}));
   }
 
-  inline auto DeviceImpl::createBuffer(vk::BufferCreateFlags flags, vk::DeviceSize size, vk::BufferUsageFlags usage) const
+  inline auto impl::Device::createBuffer(vk::BufferCreateFlags flags, vk::DeviceSize size, vk::BufferUsageFlags usage) const
   {
     return Buffer<>(BufferDeleter{magma::Device<claws::no_delete>(*this)},
                     vk::Device::createBuffer({flags, size, usage, vk::SharingMode::eConcurrent, 0, nullptr}));
