+++ include/magma/BufferView.hpp
@@ -18,7 +18,7 @@ namespace magma
   template<class Deleter = BufferViewDeleter>
   using BufferView = claws::handle<vk::BufferView, Deleter>;
 
-  inline auto DeviceImpl::createBufferView(
+  inline auto impl::Device::createBufferView(
     vk::BufferViewCreateFlags flags, claws::handle<vk::Buffer, claws::no_delete> buffer, vk::Format format, vk::DeviceSize offset, vk::DeviceSize size) const
   {
     return BufferView<>(BufferViewDeleter{magma::Device<claws::no_delete>(*this)}, vk::Device::createBufferView({flags, buffer, format, offset, size}));
