+++ include/magma/DynamicBuffer.hpp
