+++ include/magma/Framebuffer.hpp
@@ -19,7 +19,7 @@ namespace magma
   template<class Deleter = FramebufferDeleter>
   using Framebuffer = claws::handle<vk::Framebuffer, Deleter>;
 
-  inline auto DeviceImpl::createFramebuffer(
+  inline auto impl::Device::createFramebuffer(
     RenderPass<claws::no_delete> renderPass, std::vector<vk::ImageView> const &attachements, uint32_t width, uint32_t height, uint32_t layers) const
   {
     return magma::Framebuffer<>(FramebufferDeleter{magma::Device<claws::no_delete>(*this)},
