+++ include/magma/Semaphore.hpp
@@ -18,7 +18,7 @@ namespace magma
   template<class Deleter = SemaphoreDeleter>
   using Semaphore = claws::handle<vk::Semaphore, Deleter>;
 
-  inline auto DeviceImpl::createSemaphore() const
+  inline auto impl::Device::createSemaphore() const
   {
     return Semaphore<>(SemaphoreDeleter{magma::Device<claws::no_delete>(*this)}, vk::Device::createSemaphore({}));
   }
