+++ include/magma/ShaderModule.hpp
@@ -37,13 +37,13 @@ namespace magma
   using ShaderModule = claws::handle<vk::ShaderModule, Deleter>;
 
   template<class Container>
-  inline auto DeviceImpl::createShaderModule(Container const &code) const
+  inline auto impl::Device::createShaderModule(Container const &code) const
   {
     return ShaderModule<>(ShaderModuleDestructor{magma::Device<claws::no_delete>(*this)},
                           vk::Device::createShaderModule(vk::ShaderModuleCreateInfo{{}, code.size() * sizeof(uint32_t), code.data()}));
   }
 
-  inline auto DeviceImpl::createShaderModule(std::istream &input) const
+  inline auto impl::Device::createShaderModule(std::istream &input) const
   {
     auto code(inputStreamToUint32Vect(input));
